Address PR feedback for #4136
- Move drain gate to lightweight moonmind.gates namespace (stdlib-only import)
- Wire gate to live probes: collector, operator report, scoped drain metrics
- Move time-skipping fleet journeys out of integration_ci module
- Exercise artifact denial through terminal persistence with working DB
